@@ -1,24 +1,23 @@
 using System;
 
-namespace Pentago.Engine
+namespace Pentago.Engine;
+
+public class Engine : IEngine
 {
-    public class Engine : IEngine
+    private readonly string _connectionString;
+
+    public Engine(string connectionString)
     {
-        private readonly string _connectionString;
+        _connectionString = connectionString;
+    }
 
-        public Engine(string connectionString)
-        {
-            _connectionString = connectionString;
-        }
+    public Evaluation Evaluate(Board position)
+    {
+        throw new NotImplementedException();
+    }
 
-        public Evaluation Evaluate(Board position)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Move BestMove(Board position)
-        {
-            throw new NotImplementedException();
-        }
+    public Move BestMove(Board position)
+    {
+        throw new NotImplementedException();
     }
 }
