@@ -1,40 +1,39 @@
-namespace Pentago.Engine
+namespace Pentago.Engine;
+
+public class Evaluation
 {
-    public class Evaluation
+    private double? _eval;
+    private int? _winIn;
+
+    public double? Eval
     {
-        private double? _eval;
-        private int? _winIn;
-
-        public double? Eval
+        get => _eval;
+        set
         {
-            get => _eval;
-            set
-            {
-                if (value == null) return;
+            if (value == null) return;
 
-                _eval = value;
-                _winIn = null;
-            }
+            _eval = value;
+            _winIn = null;
         }
+    }
 
-        public int? WinIn
+    public int? WinIn
+    {
+        get => _winIn;
+        set
         {
-            get => _winIn;
-            set
-            {
-                if (value == null) return;
+            if (value == null) return;
 
-                _winIn = value;
-                _eval = null;
-            }
+            _winIn = value;
+            _eval = null;
         }
+    }
 
-        public override string ToString()
-        {
-            if (_eval != null) return _eval.ToString();
-            if (_winIn != null) return $"#{_winIn}";
+    public override string ToString()
+    {
+        if (_eval != null) return _eval.ToString();
+        if (_winIn != null) return $"#{_winIn}";
 
-            return "-";
-        }
+        return "-";
     }
 }
