@@ -1,63 +1,62 @@
 using System;
 
-namespace Pentago.Engine
+namespace Pentago.Engine;
+
+public struct Square
 {
-    public struct Square
+    private int _x;
+
+    private int _y;
+
+    public int X
     {
-        private int _x;
-
-        private int _y;
-
-        public int X
+        get => _x;
+        set
         {
-            get => _x;
-            set
-            {
-                if (value > 5)
-                    throw new IndexOutOfRangeException("X must be between 0 and 5 inclusive.");
+            if (value > 5)
+                throw new IndexOutOfRangeException("X must be between 0 and 5 inclusive.");
 
-                _x = value;
-            }
+            _x = value;
         }
+    }
 
-        public int Y
+    public int Y
+    {
+        get => _y;
+        set
         {
-            get => _y;
-            set
-            {
-                if (value > 5)
-                    throw new IndexOutOfRangeException("Y must be between 0 and 5 inclusive.");
+            if (value > 5)
+                throw new IndexOutOfRangeException("Y must be between 0 and 5 inclusive.");
 
-                _y = value;
-            }
+            _y = value;
         }
+    }
 
-        public Color? State { get; set; }
+    public Color? State { get; set; }
 
-        public void Rotate(bool clockwise)
+    public void Rotate(bool clockwise)
+    {
+        var x = _x;
+        var y = _y;
+        if (clockwise)
         {
-            var x = _x;
-            var y = _y;
-            if (clockwise)
-            {
-                _x += x % 3 - 1 + (x + -y) * -1;
-                _y += y % 3 - 1 + (x + y) * -1;
-            }
-            else
-            {
-                _x += x % 3 - 1 + (x + y) * -1;
-                _y += y % 3 - 1 + (-x + y) * -1;
-            }
+            _x += x % 3 - 1 + (x + -y) * -1;
+            _y += y % 3 - 1 + (x + y) * -1;
         }
-
-        public Square(int x, int y, Color? state = null)
+        else
         {
-            if (x > 5 || y > 5)
-                throw new IndexOutOfRangeException("Arguments must be between 0 and 5 inclusive.");
-            _x = x;
-            _y = y;
-
-            State = state;
+            _x += x % 3 - 1 + (x + y) * -1;
+            _y += y % 3 - 1 + (-x + y) * -1;
         }
+    }
+
+    public Square(int x, int y, Color? state = null)
+    {
+        if (x > 5 || y > 5)
+            throw new IndexOutOfRangeException("Arguments must be between 0 and 5 inclusive.");
+        _x = x;
+        _y = y;
+
+        State = state;
     }
 }
